move all css modules files to separate dir(css_modules)
