+++ src/containers/StadiumPage/StadiumPage.js
@@ -3,7 +3,7 @@ import Map from "../../components/map";
 import Stadium from "../../components/stadium-description-card";
 import { withRouter } from "react-router";
 import Swal from "sweetalert2";
-import "../../components/sweet-alert.css";
+import "../../css/sweet-alert.css";
 
 class StadiumPage extends React.Component {
   constructor(props) {
