+++ src/css/sweet-alert.css
