+++ src/css_modules/stadium-description-card.css
